% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_logo_path}
\alias{get_logo_path}
\title{Get the path to the OmicsChart logo}
\usage{
get_logo_path()
}
\value{
Character string with the path to the logo
}
\description{
Get the path to the OmicsChart logo
}
