% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_report.R
\name{pdf_report}
\alias{pdf_report}
\title{Custom RMarkdown Theme}
\usage{
pdf_report(toc = TRUE, ...)
}
\arguments{
\item{...}{Arguments passed to the output format function.}
}
\value{
A modified HTML or PDF output format with the custom theme.
}
\description{
Custom RMarkdown Theme
}
